moved utility funcs to separate module
