+++ seg_lapa/config/train.yaml
@@ -1,4 +1,5 @@
 random_seed: 0  # If None, seeds not set. If int, uses value to seed.
+logs_root_dir: './'  # Where to save logs and checkpoints.
 
 defaults:
   - dataset: lapa
