+++ seg_lapa/train.py
@@ -1,64 +1,18 @@
-import os
-from typing import Any, List, Optional
-from pathlib import Path
+from typing import Any, List
 
 import hydra
 import pytorch_lightning as pl
 import wandb
-from omegaconf import OmegaConf, DictConfig
+from omegaconf import DictConfig
 
 from seg_lapa import metrics
 from seg_lapa.config_parse import train_conf
 from seg_lapa.config_parse.train_conf import TrainConf
 from seg_lapa.loss_func import CrossEntropy2D
 from seg_lapa.utils.path_check import get_project_root
-from seg_lapa.callbacks.log_media import Mode, LogMediaQueue, LogMedia
-
-LOGS_DIR = "logs"
-
-
-def is_rank_zero():
-    local_rank = int(os.environ.get("LOCAL_RANK", 0))
-    node_rank = int(os.environ.get("NODE_RANK", 0))
-    if local_rank == 0 and node_rank == 0:
-        return True
-
-    return False
-
-
-def generate_run_id(cfg: DictConfig):
-    # Set the run ID: Read from config if resuming training, else generate unique id
-    # TODO: read from cfg if resuming training - get from config dataclass! add method to resume training section.
-    run_id = wandb.util.generate_id()
-    return run_id
-
-
-def create_log_dir(cfg: DictConfig, run_id: str) -> Optional[Path]:
-    """Each run's log dir will have same name as wandb runid"""
-    log_root_dir = get_project_root() / LOGS_DIR
-
-    if is_rank_zero():
-        log_dir = log_root_dir / run_id
-        log_dir.mkdir(parents=True, exist_ok=True)
-
-        # Save the input config file to logs dir
-        OmegaConf.save(cfg, log_dir / "train.yaml")
-    else:
-        return log_root_dir / "None"
-
-    return log_dir
-
-
-def fix_seeds(random_seed: Optional[int]) -> None:
-    """Fix seeds for reproducibility.
-    Ref:
-        https://pytorch.org/docs/stable/notes/randomness.html
-
-    Args:
-        random_seed: If None, seeds not set. If int, uses value to seed.
-    """
-    if random_seed is not None:
-        pl.seed_everything(random_seed)
+from seg_lapa.callbacks.log_media import Mode, LogMediaQueue
+from seg_lapa.utils.utils import is_rank_zero
+from seg_lapa.utils import utils
 
 
 class DeeplabV3plus(pl.LightningModule):
@@ -177,11 +131,11 @@ def main(cfg: DictConfig):
     if is_rank_zero():
         print("\nResolved Dataclass:\n", config, "\n")
 
-    fix_seeds(config.random_seed)
-    run_id = generate_run_id(cfg)
-    log_dir = create_log_dir(cfg, run_id)
+    utils.fix_seeds(config.random_seed)
+    run_id = utils.generate_run_id(cfg)
+    log_dir = utils.create_log_dir(cfg, run_id)
 
-    wb_logger = config.logger.get_logger(cfg, run_id, get_project_root())
+    wb_logger = config.logger.get_logger(cfg, run_id, get_project_root() / utils.LOGS_DIR)
     callbacks = config.callbacks.get_callbacks_list(log_dir)
     trainer = config.trainer.get_trainer(wb_logger, callbacks, get_project_root())
     model = DeeplabV3plus(config)
